+++ src/components/index.ts
@@ -1,7 +1,8 @@
 // atoms
 export { default as Button } from "./Button";
+export { default as Input } from "./Input";
 // patterns
 export { default as Layout } from "./Layout";
 export { default as Navigation } from "./Navigation";
-export { default as Main, MainWrapper as MainWrapper } from "./Main";
+export { default as Main } from "./Main";
 export { default as Footer } from "./Footer";